move single function to jwk.go, delete headers.go
